Fixed chapter, added user_roles library

+++ application/controllers/chapter.php
@@ -43,6 +43,39 @@ class Chapter extends CI_COntroller {
   
   public function display(){
     
+    //limit display of chapters only to authorized
+    if( !$this->user_roles->is_authorized( array('admin_chapter_view') ) ):
+        $list['error_msg'] = "You are not authorized to view Chapters.";
+        $list['main_content'] = "message_view";
+        $this->load->view('includes/template', $list);
+        return;
+    endif;
+    
+    //get ajax request for tags
+	if( $this->input->is_ajax_request() ):
+      $data = $this->input->post('data');
+      if( $data['task'] == 'hadith-book' ):
+      
+        $this->load->model('book_model');
+      
+        //get respected books for hadith book id
+        $books = $this->book_model->get_all_books( $data['hadith_book_id'] );
+        
+        //html for options
+        $book_html='';
+        foreach( $books as $book ):
+            $book_html .= '<option value="'.$book->book_id.'">'.$book->book_title_en.'</option>';
+        endforeach;
+        
+        
+        $data = new stdClass();
+        $data->book_list = $book_html;
+        
+        echo json_encode( $data );
+			
+		return;
+      endif;
+    endif;
     $this->load->helper('form');
     $this->load->model('chapter_model');
     $list['ahadith'] = $this->chapter_model->get_all_chapters();
@@ -58,9 +91,17 @@ class Chapter extends CI_COntroller {
   
   public function add(){
   
+    //limit updating of chapter only to authorized
+    if( !$this->user_roles->is_authorized( array('admin_chapter_add') ) ):
+        $list['error_msg'] = "You are not authorized to Add Chapter.";
+        $list['main_content'] = "message_view";
+        $this->load->view('includes/template', $list);
+        return;
+    endif;
+  
     $this->load->model('book_model');
-    $list['books'] = $this->book_model->get_all_books();
     $list['hadith_books'] = $this->book_model->get_all_hadith_books();
+    $list['books'] = $this->book_model->get_all_books( $list['hadith_books'][0]->hadith_book_id );
     $list['main_content'] = 'admin/add_chapter_view';
     
     $this->load->helper('form');
@@ -107,6 +148,14 @@ class Chapter extends CI_COntroller {
   
   public function update($chapter_id){
 
+    //limit updating of chapter only to authorized
+    if( !$this->user_roles->is_authorized( array('admin_chapter_edit') ) ):
+        $list['error_msg'] = "You are not authorized to modify Chapter.";
+        $list['main_content'] = "message_view";
+        $this->load->view('includes/template', $list);
+        return;
+    endif;
+  
     $this->load->model('chapter_model');
     
     if( $this->chapter_model->get_chapter_by_id($chapter_id) == FALSE ):
@@ -129,8 +178,8 @@ class Chapter extends CI_COntroller {
     
   
     $this->load->model('book_model');
-    $list['books'] = $this->book_model->get_all_books();
     $list['hadith_books'] = $this->book_model->get_all_hadith_books();
+    $list['books'] = $this->book_model->get_all_books( $list['hadith_books'][0]->hadith_book_id );
     $list['chapter_id'] = $chapter_id;
     $list['chapter'] =  $this->chapter_model->get_chapter_by_id($chapter_id);
     $list['main_content'] = 'admin/update_chapter_view';
@@ -164,8 +213,18 @@ class Chapter extends CI_COntroller {
    *@return none
    */
   
-  public function delete( $chapter_id){
+  public function delete( $chapter_id=''){
 
+    //limit deleting of employee appointments only to authorized
+    if( !$this->user_roles->is_authorized( array('admin_chapter_delete') ) ):
+        $list['error_msg'] = "You are not authorized to Delete Chapters.";
+        $list['main_content'] = "message_view";
+        $this->load->view('includes/template', $list);
+        return;
+        
+    endif;
+  
+  
     $this->load->helper('url');
     $this->load->model('chapter_model');
     $this->chapter_model->delete_chapter( $chapter_id );
